Added tetron rifle & gun refactor
