+++ Stealth Game/Assets/Materials/Effects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: afde1c41f21a15e44aa48883f19eeea3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
