+++ Stealth Game/Assets/Materials/Effects/ConcreteBulletHole.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9e1f6f1413210f642866f3fc59ea7c58
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
