+++ Stealth Game/Assets/Materials/Effects/ImpactLightning.mat.meta
