+++ Stealth Game/Assets/Materials/Effects/ImpactSmoke.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 364e9a1eee6572f4f8e07e4c50c78b3c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
