+++ Stealth Game/Assets/Materials/Effects/RopeLightning.mat
