+++ Stealth Game/Assets/Materials/Effects/Taser Wires.mat
