+++ Stealth Game/Assets/Meshes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e866f07dbe825ad45b44d1a01432e697
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
