+++ Stealth Game/Assets/Models/Tetron Rifle.fbx
