+++ Stealth Game/Assets/Models/Tetron Rifle.fbx.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f403abd52641ceb468b1a0c93e779ab5
+guid: 4e00238f77cc6d348963ce091faaea11
 ModelImporter:
   serializedVersion: 22200
   internalIDToNameTable: []
