+++ Stealth Game/Assets/Prefabs/ImpactEffect.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: b1f1b412d57aceb41a5c8ea8a49d4208
-NativeFormatImporter:
+guid: f03506c64edd8674283916e54510bed4
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
