+++ Stealth Game/Assets/Prefabs/Tetron Rifle.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5a482a099bad3b24d9af84c405bff965
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
