+++ Stealth Game/Assets/Scripts/DestroyAfterTime.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+public class DestroyAfterTime : MonoBehaviour
+{
+    public float delay;
+
+    private void Start()
+    {
+        Destroy(gameObject, delay);
+    }
+}
