+++ Stealth Game/Assets/Scripts/DestroyAfterTime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a09032a5569adcf4caddb2b85a8b9d7e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
