+++ Stealth Game/Assets/Scripts/Guns/HitscanGun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c236fa333e6feb42970e6e95b97a51f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
