+++ Stealth Game/Assets/Scripts/Guns/ProjectileGun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e02d53eec71f8b40bd00774d690ba2b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
