+++ Stealth Game/Assets/Scripts/Guns/Tetron.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4dac49b879e99724cb62fc640237bdc7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
